追加:courseの詳細ページにルート検索結果の追加 #12

+++ db/migrate/20210701085213_change_courses_columns.rb
@@ -2,7 +2,7 @@ class ChangeCoursesColumns < ActiveRecord::Migration[6.0]
   def up
     change_column :courses, :starting_point_lat, :decimal, precision: 10, scale: 7
     change_column :courses, :starting_point_lng, :decimal, precision: 10, scale: 7
-    change_column :courses, :ascent_time, :float
-    change_column :courses, :descent_time, :float
+    change_column :courses, :ascent_time, :integer
+    change_column :courses, :descent_time, :integer
   end
 end
